+++ packages/@weresk/router/src/utils/index.ts
@@ -1,4 +1,5 @@
 export * from "./getDocumentPath";
+export * from "./getLinkType";
 export * from "./getTypesWithSameRoute";
 export * from "./isUniqueSlugFunction";
 export * from "./normalizeLink";