+++ packages/@weresk/locales/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @weresk/locales
 
+## 0.0.4
+
+### Patch Changes
+
+- Making 'lang' property optional for LocaleProps
+
 ## 0.0.3
 
 ### Patch Changes