+++ packages/@weresk/router/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @weresk/router
 
+## 0.0.2
+
+### Patch Changes
+
+- Adding LinkWrapper and ContactLink components
+
 ## 0.0.1
 
 ### Patch Changes